Messing around with GUI
Tried to do stuff with GUI... didn't end up too well.
Cat2 - is the dropped attempt at making it text based (didn't finish it)
Game - was suppose to be the general game
Frame - The majority of the work was done here. Managed to make a window appear in the center of the computer screen, have a button where when you press it, the screen actually changes from yellow to the default color, gray