More work on Frame
Tried to do more stuff on Frame where once you press the "Play Game" button, it will change to a new JPanel which has a button on it where you can click on that and it will change the JPanel too. But, the button doesn't show up in the first place so this is a NO GO overall.